@@ -15,9 +15,9 @@
  */
 package org.springframework.samples.petclinic.model;
 
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import javax.validation.constraints.NotEmpty;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.validation.constraints.NotBlank;
 
 /**
  * Simple JavaBean domain object representing an person.
@@ -28,11 +28,11 @@ import javax.validation.constraints.NotEmpty;
 public class Person extends BaseEntity {
 
 	@Column(name = "first_name")
-	@NotEmpty
+	@NotBlank
 	private String firstName;
 
 	@Column(name = "last_name")
-	@NotEmpty
+	@NotBlank
 	private String lastName;
 
 	public String getFirstName() {
