@@ -16,6 +16,9 @@
 
 package org.springframework.samples.petclinic.owner;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -25,12 +28,10 @@ import java.util.Locale;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledInNativeImage;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.BDDMockito.given;
 
 /**
  * Test class for {@link PetTypeFormatter}
@@ -38,10 +39,11 @@ import static org.mockito.BDDMockito.given;
  * @author Colin But
  */
 @ExtendWith(MockitoExtension.class)
+@DisabledInNativeImage
 class PetTypeFormatterTests {
 
 	@Mock
-	private PetRepository pets;
+	private OwnerRepository pets;
 
 	private PetTypeFormatter petTypeFormatter;
 
