@@ -16,36 +16,31 @@
 
 package org.springframework.samples.petclinic.system;
 
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.forwardedUrl;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
+import org.junit.jupiter.api.Test;
 
 /**
  * Test class for {@link CrashController}
  *
  * @author Colin But
+ * @author Alex Lutz
  */
-// Waiting https://github.com/spring-projects/spring-boot/issues/5574
-@Disabled
-@WebMvcTest(controllers = CrashController.class)
+// Waiting https://github.com/spring-projects/spring-boot/issues/5574 ..good
+// luck ((plain(st) UNIT test)! :)
 class CrashControllerTests {
 
-	@Autowired
-	private MockMvc mockMvc;
+	CrashController testee = new CrashController();
 
 	@Test
 	void testTriggerException() throws Exception {
-		mockMvc.perform(get("/oups")).andExpect(view().name("exception"))
-				.andExpect(model().attributeExists("exception")).andExpect(forwardedUrl("exception"))
-				.andExpect(status().isOk());
+		RuntimeException thrown = assertThrows(RuntimeException.class, () -> {
+			testee.triggerException();
+		});
+
+		assertEquals("Expected: controller used to showcase what happens when an exception is thrown",
+				thrown.getMessage());
 	}
 
 }
